fix: export all types in types folder
There was an issue with the rollup config in which only two files:
- src/index.ts
- src/testing/index.ts
were included as input for rollup. Rollup only traced references from these two files to generate the output bundle.

In the types folder, there were exported types that weren't referred to by either of the two inputs. These types weren't bundled into the output.

This PR aims to include all files in the types folder into the bundle output.