feat: Readonly calls from app client now don't invoke signing feat: AlgoKitComposer.simulate() now has a `skipSignatures` option to opt-out of signing for the simulate call feat: Added a `defaultSigner` to `AppClient` and `AppFactory` feat: `AppClient` now has an `algorand` property so you don't need to pass `AlgorandClient` around everywhere if you have a client feat: `AppFactory` now has an `algorand` property so you don't need to pass `AlgorandClient` around everywhere if you have a factory feat: `AppFactory` now has `appName` and `appSpec` properties so if you want this information it's exposed rather than private feat: `appFactory.deploy` now allows you to override `appName` so if you want to create multiple contracts from a factory with different names you can
BREAKING CHANGE: AppClient.compile (and appClient.compile, and appFactory.compile) return `compiledApproval` and `compiledClear` again
This reverted the last commit after a reflection that the more explicit naming was actually confusing since it meant
the return of this method different from the return from create/update calls and consistency was better. The fact there is now jsdoc
comments means that the property names from the compile method are able to be quickly understood.