feat!: replace AlgorandClientInterface with AlgorandClient (#387)
See the discussion in
https://github.com/algorandfoundation/algokit-utils-ts/pull/365 for
context. Essentially by continuing to use the Interface any additional
feature to AlgorandClient becomes a breaking change. The only reason for
using the interface in the first place was for a rollup error, but this
was a false positive and can be fixed by using the type import as shown
in this commit